start working with a db

+++ src/db/subscriptions.rs
@@ -1,5 +1,5 @@
 use crate::models::subscription::Subscription;
 
-pub async fn find_subscriptions_by_id(_id: i32) -> Option<Subscription> {
+pub async fn find_subscriptions_by_id(_id: i64) -> Option<Subscription> {
     Option::from(Subscription::new_subscription())
 }
